Moved show panel closer to the plots.
