+++ react/features/variant/VariantPlots.js
@@ -10,7 +10,7 @@ export const VariantPlots = () => {
 	const options = useSelector(state => state.data.options)    
 
     return (
-	    <div>
+	    <div style={{marginTop: "20px"}}>
 	    {
 	    	options.barmap == 'map' && options.bbreg == 'region' ? <VariantPlotMap /> : <VariantPlotCohort /> 
 	    }
